Update package structure
