+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup
 
 setup(name="simulink_gym",
-      version="0.3.0",
+      version="0.4.0",
       author="Johannes Brust",
       packages=["simulink_gym"]
       )
