+++ simulink_gym/__init__.py
@@ -1,6 +1,4 @@
+from .environment import Environment, param_block
 from .actions import Action, Actions
 from .observations import Observation, Observations
-from .environment import Environment, CommSocket, param_block
 from gym import logger
-
-__all__ = ['Action', 'Actions', 'Observation', 'Observations', 'Environment', 'CommSocket', 'param_block', 'logger']
