+++ simulink_gym/utils/__init__.py
@@ -1,0 +1,1 @@
+from .comm_socket import CommSocket
